+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "3.10.0-rc.5",
+  sdkVersion: "3.10.0-rc.6",
   genVersion: "2.656.5",
-  userAgent: "speakeasy-sdk/typescript 3.10.0-rc.5 2.656.5 2.0 @orq-ai/node",
+  userAgent: "speakeasy-sdk/typescript 3.10.0-rc.6 2.656.5 2.0 @orq-ai/node",
 } as const;