+++ packages/orq-rc/RELEASES.md
@@ -2489,3 +2489,13 @@ Based on:
 - [typescript v3.10.0-rc.5] packages/orq-rc
 ### Releases
 - [NPM v3.10.0-rc.5] https://www.npmjs.com/package/@orq-ai/node/v/3.10.0-rc.5 - packages/orq-rc
+
+## 2025-07-11 12:54:05
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.581.0 (2.656.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.10.0-rc.6] packages/orq-rc
+### Releases
+- [NPM v3.10.0-rc.6] https://www.npmjs.com/package/@orq-ai/node/v/3.10.0-rc.6 - packages/orq-rc